# Speed up release notes API by narrowing sooner.